requirements.txt
# Core FastAPI
fastapi
uvicorn[standard]
python-multipart
python-dotenv

# Excel processing
openpyxl
Pillow

# Authentication
clerk-sdk-python

# HTTP client
httpx
